% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_index2.R
\name{alphaestimator2}
\alias{alphaestimator2}
\title{alphaestimator2}
\usage{
alphaestimator2(path0)
}
\arguments{
\item{path0}{(Vector of univariate observations)}
}
\value{
An integer with the tail - index estimate
}
\description{
This function computes the (tail)-index.
It returns the mean estimate from the Pickands, Hill and DeHaan estimators
at 96th empirical quantile.
It uses the fExtreme code.
}
\examples{
sample   <- abs( arima.sim(n = 8000, list(ar=0.5, ma=0), rand.gen=function(n) rt(n,df=4) ) )
alphaestimator2(sample)
}
