% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ARCH2m}
\alias{ARCH2m}
\title{ARCH2m}
\usage{
ARCH2m(
  n0,
  eta = 2 * 10^{
     -5
 },
  lambda = 0.5
)
}
\arguments{
\item{n0}{(Integer with trajectory length)}

\item{eta}{(Real value)}

\item{lambda}{(Real value)}
}
\value{
(Vector of length n0 sampled from the SRE model)
}
\description{
Function to sample from the model
X= AX + B     with X = (eta + lambda X) Z^2, Z iid Gaussian.
For the default values, this model has index of regular variation alpha = 1
and extremal index theta = 0.727
}
\examples{
path <- ARCH2m(1000)
plot.ts(path)
}
