% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ARCHm}
\alias{ARCHm}
\title{ARCHm}
\usage{
ARCHm(n0)
}
\arguments{
\item{n0}{(Integer with trajectory length)}
}
\value{
(Vector of length n0 sampled from the SRE model)
}
\description{
Function to sample from the model
X  = AX + B     with log A ~ N-0.5, B ~ Unif(0,1)
The model has index of regular variation alpha = 1
and extremal index theta = 0.2792
}
\examples{
path <- ARCHm(1000)
plot.ts(path)
}
