% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ARm}
\alias{ARm}
\title{AR model}
\usage{
ARm(n0, par0 = 0.7, Z.gen = function(n) rt(n, df = 1))
}
\arguments{
\item{n0}{(Integer with trajectory length)}

\item{par0}{(Value in (0,1))}

\item{Z.gen}{(Function of a trajectory length n to sample from the noise)}
}
\value{
(Vector of length n0 sampled from the SRE model)
}
\description{
Function to sample from the model
X = (par0)X + Z.
For the default values, this model has index of regular variation alpha identical to Z
and extremal index 1-(par0)^alpha
}
\examples{
path  <- ARm(1000,0.7)
plot.ts(path)
}
