% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremogram.R
\name{ext}
\alias{ext}
\title{Extremogram auxiliar function}
\usage{
ext(sample, maxlag = 10, q)
}
\arguments{
\item{sample}{A vector}

\item{maxlag}{An integer}

\item{q}{A quantile value in $(0,1)$}
}
\value{
A vector with the extremogram output
}
\description{
Extremogram auxiliar function
}
