% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremogram.R
\name{extremogram}
\alias{extremogram}
\title{Empirical temporal extremogram}
\usage{
extremogram(sample, maxlag = 45, q = 0.95, plot = T)
}
\arguments{
\item{sample}{A numerical vector to compute the extremogram}

\item{maxlag}{An integer with the}

\item{q}{A quantile level in (0,1).}

\item{plot}{A logical value to plot the results}
}
\value{
A vector with the extremogram values
}
\description{
Empirical temporal extremogram
}
\details{
The temporal extremogram \eqn{\chi_t} is defined by
\deqn{\chi_t = \lim_{t \to \infty} \mathbb{P}(|X_t| > x | |X_0| > x).}
Its empirical version, computes the average number of exceedances of the q-th order statistic.
As a baseline, the extremogram takes the value of $1-q$ at independent lags.
}
\examples{
pre    <- tsExtremes::rainfall ## load rainfall data
sample <- na.omit(sample)      ## removes columns with missing values
sample <- tsExtremes::rmax(pre[pre$SEASON=='FALL' , 5:7]) ## computes vector-wise maxima
tsExtremes::extremogram(sample)

}
