% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functionals.R
\name{functionalCP}
\alias{functionalCP}
\title{functionalCP}
\usage{
functionalCP(
  path0,
  alpha0,
  klim0 = 100,
  n0 = length(path0),
  p0 = "alpha",
  cluster_func = function(block, p0, alpha0) max(abs(block)^alpha0),
  plot = F
)
}
\arguments{
\item{path0}{(Matrix with the sample trajectory of a time series)}

\item{alpha0}{(Integer with an estimate of the tail index)}

\item{klim0}{(Integer with the largest order statistic k to consider, or a vector with values k to consider)}

\item{n0}{(Integer with the path length)}

\item{cluster_func}{(A function to evaluate on extremal blocks. As a default we provide the functional for computing the extremal index)}

\item{plot}{(F or T if the plot with estimates as a function of k must be shown)}
}
\value{
data frame with estimates of the the cluster functional estimates values of k and block lengths b used for inference
An estimate of the asymptotic variance is also provided using cluster process inference.
}
\description{
Estimates of the cluster statistics and based on the alpha cluster Process
}
\examples{
path  <- ARCHm(20000)
alpha <- 1/alphaestimator(path,k1=1200)$xi ## The real value should be one
two options are available for computing the extremal index
ei  <- functionalCP(path,1,100,plot=T)
ei2 <- eiCP(path,1,100,plot=T)
}
