% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_index.R
\name{hillestimator}
\alias{hillestimator}
\title{Compute (tail)-index}
\usage{
hillestimator(path, k1 = NULL, plot = F)
}
\arguments{
\item{path}{(vector of univariate observations)}

\item{k1}{(vector of order statistics to compute Hill estimates).
It uses k1 = n*0.05 otherwise}

\item{plot}{(T or F if a plot should be shown).}
}
\value{
A data.frame of hill estimates as a function of k.
}
\description{
This function computes Hill (tail)-index estimate of a stationary time series.
}
\examples{
h <- hillestimator(rainfall$LANVEOC[rainfall$SEASON=='SPRING'], 1:200)
h <- hillestimator(rainfall$LANVEOC[rainfall$SEASON=='SPRING'], plot=T)
}
