% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_functional.R
\name{piCP}
\alias{piCP}
\title{piCP}
\usage{
piCP(path0, alpha0, klim0 = 100, n0 = length(path0), plot = F)
}
\arguments{
\item{path0}{(Matrix with the sample trajectory of a time series)}

\item{alpha0}{(Integer with an estimate of the tail index)}

\item{klim0}{(Integer with the largest order statistic k to consider, or a vector with values k to consider)}

\item{n0}{(Integer with the path length)}

\item{plot}{(F or T if the plot with estimates as a function of k must be shown)}
}
\value{
data frame with estimates of the the cluster lengths pi_j with values of k and block lengths b used for inference
An estimate of the asymptotic variance is also provided using cluster process inference.
}
\description{
Estimates of the cluster legths and the extrema index based on the alpha cluster Process
}
\examples{
path  <- ARCHm(20000)
alpha <- 1/alphaestimator(path,k1=1200)$xi ## The real value should be one
pi    <- piCP(path,alpha,plot=T)
}
