% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Daily rainfall data set from a network of weather stations in France}
\format{
A data frame with 10 columns.
\describe{
\item{y.date}{Date}
\item{BREST}{Precipitation amounts in Brest (oceanic weather)}
\item{LANVEOC}{Precipitation amounts in Lanveoc (oceanic weather)}
\item{QUIMPER}{Precipitation amounts in Quimper (oceanic weather)}
\item{BORMES}{Precipitation amounts in Bormes (mediterranean weather)}
\item{LE.LUC}{Precipitation amounts in Le Luc (mediterranean weather)}
\item{HYERES}{Precipitation amounts in Hyeres (mediterranean weather)}
\item{NANCY}{Precipitation amounts in Nancy (continental weather)}
\item{METZ}{Precipitation amounts in Metz (continental weather)}
\item{ROVILLE}{Precipitation amounts in Roville (continental weather)}
\item{SEASON}{Season}
}
}
\source{
Méteo France.
}
\usage{
rainfall
}
\description{
A data set with precipitation amounts in mm from nine weather stations.
}
\keyword{datasets}
