% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremogram.R
\name{rmax}
\alias{rmax}
\title{Row maxima}
\usage{
rmax(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
A numerical vector with the same number of columns as \code{mat}
}
\description{
\code{rmax()} computes the maxima row-wise
}
\examples{
x <- rbind( c(1,2), c(1,1) )
rmax(x)
}
